requirements.txt
opencv-contrib-python
numpy
pyautogui
pillow
pyttsx3
ctypes
pyHook
win32gui
logging
